cleaning up and adding tests
