+++ backend/messaging-channel/src/main/kotlin/io/featurehub/messaging/service/FeatureMessagingCloudEventPublisherImpl.kt
@@ -1,113 +1,78 @@
 package io.featurehub.messaging.service
 
 import io.cloudevents.core.builder.CloudEventBuilder
+import io.featurehub.db.api.RolloutStrategyUpdate
 import io.featurehub.events.CloudEventDynamicPublisherRegistry
 import io.featurehub.events.CloudEventPublisher
+import io.featurehub.events.DynamicCloudEventDestination
+import io.featurehub.events.DynamicCloudEventDestinationMapper
 import io.featurehub.lifecycle.LifecycleListener
-import io.featurehub.lifecycle.LifecycleListeners
 import io.featurehub.lifecycle.LifecyclePriority
-import io.featurehub.messaging.MessagingConfig
 import io.featurehub.messaging.model.FeatureMessagingUpdate
+import io.featurehub.messaging.model.MessagingFeatureValueUpdate
+import io.featurehub.messaging.model.MessagingLockUpdate
+import io.featurehub.messaging.model.MessagingRetiredUpdate
+import io.featurehub.messaging.model.MessagingRolloutStrategy
+import io.featurehub.messaging.model.MessagingRolloutStrategyAttribute
+import io.featurehub.messaging.model.MessagingStrategiesReorder
+import io.featurehub.messaging.model.MessagingStrategyUpdate
+import io.featurehub.messaging.model.StrategyUpdateType
+import io.featurehub.mr.model.FeatureVersion
+import io.featurehub.mr.model.RolloutStrategy
+import io.featurehub.mr.model.RolloutStrategyAttribute
+import io.featurehub.utils.ExecutorSupplier
 import io.featurehub.utils.FallbackPropertyConfig
 import jakarta.inject.Inject
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.net.URI
 import java.time.OffsetDateTime
+import java.time.ZoneOffset
 import java.util.*
+import java.util.concurrent.ExecutorService
 
-/**
- * The purpose of this type is to associate a cloud event type with configuration provided by the person who is running
- * FeatureHub. It goes and tells the system using the dynamic publishers to register a listener for a cloud event of type X
- * against a publisher of type Y. Extra headers can be provided for some transport mechanisms (currently only http/https)
- */
-class MappedSupportedConfig(val cloudEventType: String, private val prefix: String, private val configInfix: String, dynamicPublisherRegistry: CloudEventDynamicPublisherRegistry) {
-  // used to determine where to send this
-  val destination = FallbackPropertyConfig.getMandatoryConfig("${prefix}.destination")
-
-  // used to automatically add to each outgoing CE message
-  val headers =
-    FallbackPropertyConfig.getConfig("${prefix}.headers", "").split(",").map { it.trim() }.filter { it.isNotEmpty() }.map {
-      if (it.contains("=")) {
-        val splits = it.split("=")
-        splits[0] to splits[1]
-      } else { // allow for secrets
-        it to FallbackPropertyConfig.getConfig("${prefix}.headers.${it}", "")
-      }
-    }.toMap()
-
-  init {
-    // this will create a new publisher based on config for this type
-    dynamicPublisherRegistry.requireDynamicPublisher(destination, headers, cloudEventType)
-
-    log.info("registered new dynamic publisher $configInfix for destination $destination for CE-type of $cloudEventType (headers $headers)")
-  }
-
-  // augumented into the body of the message
-  fun additionalProperties(sourceWebhookMap: Map<String,String>): Map<String, String> {
-    return sourceWebhookMap
-      .filter { it.key.startsWith(configInfix) }
-      .filter { it.key != "$prefix.enabled" }
-      .map { it.key to it.value }.toMap()
-  }
-
-  companion object {
-    private val log: Logger = LoggerFactory.getLogger(MappedSupportedConfig::class.java)
-
-    fun infix(infix: String, dynamicPublisherRegistry: CloudEventDynamicPublisherRegistry): String? {
-      if (FallbackPropertyConfig.getConfig("webhook.${infix}.enabled", "false") == "true") {
-        FallbackPropertyConfig.getConfig("webhook.${infix}.destination")?.let {
-          if (dynamicPublisherRegistry.confirmDynamicPublisherExists(it)) {
-            return "webhook.${infix}"
-          }
-        }
-
-        FallbackPropertyConfig.getConfig("webhook.default.destination")?.let {
-          if (dynamicPublisherRegistry.confirmDynamicPublisherExists(it)) {
-            return "webhook.default"
-            }
-        }
-      }
-
-      return null
-    }
-  }
-}
 
 @LifecyclePriority(priority = 12) // this is after all the dynamic registries have registered (they register at priority 5)
 class FeatureMessagingCloudEventInitializer @Inject constructor(publisher: FeatureMessagingCloudEventPublisher,
                                                           dynamicPublisherRegistry: CloudEventDynamicPublisherRegistry) : LifecycleListener {
   init {
     val hooks = mapOf(Pair("integration.slack", "integration/slack-v1")).map {
-      val prefix = MappedSupportedConfig.infix(it.key, dynamicPublisherRegistry)
-      if (prefix != null) MappedSupportedConfig(it.value, prefix, it.key, dynamicPublisherRegistry) else null
+      val prefix = DynamicCloudEventDestinationMapper.infix(it.key, dynamicPublisherRegistry)
+      if (prefix != null) DynamicCloudEventDestinationMapper(it.value, prefix, it.key, dynamicPublisherRegistry) else null
     }.filterNotNull()
 
     publisher.setHooks(hooks)
   }
 }
 
-class FeatureMessagingCloudEventPublisherImpl @Inject constructor(
-  private val cloudEventPublisher: CloudEventPublisher, private val messagingConfig: MessagingConfig
+open class FeatureMessagingCloudEventPublisherImpl @Inject constructor(
+  private val cloudEventPublisher: CloudEventPublisher,
+  private val supplier: ExecutorSupplier
 ) : FeatureMessagingCloudEventPublisher {
-  private val log = LoggerFactory.getLogger(FeatureMessagingCloudEventPublisherImpl::class.java)
-  private val hooks = mutableListOf<MappedSupportedConfig>();
+  private val log: Logger = LoggerFactory.getLogger(FeatureMessagingCloudEventPublisherImpl::class.java)
+  private val hooks = mutableListOf<DynamicCloudEventDestination>();
+  private var executor: ExecutorService? = null
 
-  override fun setHooks(hooks: List<MappedSupportedConfig>) {
+  override fun setHooks(hooks: List<DynamicCloudEventDestination>) {
     this.hooks.clear()
     this.hooks.addAll(hooks)
+
+    if (hooks.isNotEmpty()) {
+      executor = supplier.executorService(FallbackPropertyConfig.getConfig("messaging.publish.thread-pool", "4").toInt())
+    }
   }
 
-  override fun publishFeatureMessagingUpdate(
-    webhookEnvironmentInfo: Map<String, String>,
-    converterHandler: () -> FeatureMessagingUpdate
-  ) {
+  override val isEnabled: Boolean
+    get() = hooks.isNotEmpty()
+
+  override fun publish(featureMessagingParameter: FeatureMessagingParameter) {
     if (hooks.isEmpty()) return // skip as we have no enabled hooks
 
-    messagingConfig.executor?.let {
+    executor?.let {
       it.submit {
         try {
-          val featureMessagingUpdate = converterHandler()
+          val featureMessagingUpdate = toFeatureMessagingUpdate(featureMessagingParameter)
+          val webhookEnvironmentInfo = featureMessagingParameter.featureValue.environment.webhookEnvironmentInfo ?: mapOf()
           val messageId = UUID.randomUUID().toString()
 
           hooks.forEach { hook ->
@@ -130,5 +95,120 @@ class FeatureMessagingCloudEventPublisherImpl @Inject constructor(
         }
       }
     }
+  }
+
+  open fun toFeatureMessagingUpdate(
+    featureMessagingParameter: FeatureMessagingParameter
+  ): FeatureMessagingUpdate {
+    try {
+      val featureValue = featureMessagingParameter.featureValue
+      val environment = featureValue.environment
+      val parentApplication = environment.parentApplication
+      val portfolio = parentApplication.portfolio
+      return FeatureMessagingUpdate()
+        .featureKey(featureValue.feature.key)
+        .featureId(featureValue.feature.id)
+        .featureName(featureValue.feature.name)
+        .featureValueId(featureValue.id!!)
+        .environmentName(environment.name)
+        .environmentId(environment.id)
+        .whoUpdatedId(featureValue.whoUpdated.id ?: UUID(0,0))
+        .whoUpdated(featureValue.whoUpdated.name ?: "")
+        .whenUpdated(featureValue.whenUpdated.atOffset(ZoneOffset.UTC))
+        .applicationId(parentApplication.id)
+        .appName(parentApplication.name)
+        .portfolioId(portfolio.id)
+        .portfolioName(portfolio.name)
+        .organizationId(portfolio.organization.id)
+        .orgName(portfolio.organization.name)
+        .version(FeatureVersion().curr(featureMessagingParameter.versionUpdate.updated!!).prev(featureMessagingParameter.versionUpdate.previous))
+        .featureValueType(featureValue.feature.valueType)
+        .desc(featureValue.feature.description)
+        .link(featureValue.feature.link)
+        .let {
+          val defaultValueUpdate = featureMessagingParameter.defaultValueUpdate
+          val defaultValueUpdated = defaultValueUpdate.updated
+          val defaultValuePrevious = defaultValueUpdate.previous
+          if (defaultValueUpdate.hasChanged) it.featureValueUpdated(
+            MessagingFeatureValueUpdate()
+              .updated(defaultValueUpdated)
+              .previous(defaultValuePrevious)
+          )
+          else it
+        }
+        .let {
+          val lockUpdated = featureMessagingParameter.lockUpdate.updated
+          val lockPrevious = featureMessagingParameter.lockUpdate.previous
+          if (featureMessagingParameter.lockUpdate.hasChanged)
+            it.lockUpdated(
+              MessagingLockUpdate()
+              .updated(lockUpdated)
+              .previous(lockPrevious)
+            ) else it
+        }
+        .let {
+          val retiredUpdated = featureMessagingParameter.retiredUpdate.updated
+          val retiredPrevious = featureMessagingParameter.retiredUpdate.previous
+          if (featureMessagingParameter.retiredUpdate.hasChanged)
+            it.retiredUpdated(
+              MessagingRetiredUpdate()
+              .updated(retiredUpdated)
+              .previous(retiredPrevious)
+            ) else it
+        }
+        .let {
+          val messagingStrategiesReorder = MessagingStrategiesReorder()
+          val strategyUpdates = featureMessagingParameter.strategyUpdates
+          if (strategyUpdates.hasChanged) {
+            if (strategyUpdates.updated.isNotEmpty())
+              it.strategiesUpdated(
+                strategyUpdates.updated.map { rolloutStrategyUpdate -> toMessagingStrategyUpdate(rolloutStrategyUpdate) })
+
+            if (strategyUpdates.reordered.isNotEmpty())
+              it.strategiesReordered(
+                messagingStrategiesReorder.reordered(
+                  strategyUpdates.reordered.map { rolloutStrategy -> toMessagingRolloutStrategy(rolloutStrategy) }
+                ))
+
+            if (strategyUpdates.previous.isNotEmpty())
+              it.strategiesReordered(
+                messagingStrategiesReorder.previous(
+                  strategyUpdates.previous.map { rolloutStrategy -> toMessagingRolloutStrategy(rolloutStrategy) }
+                ))
+          }
+          it
+        }
+    } catch (e: Exception) {
+      log.error("Unable to convert feature messaging parameter {}",featureMessagingParameter, e)
+      throw(e)
+    }
+  }
+
+  private fun toMessagingStrategyUpdate(rolloutStrategyUpdate: RolloutStrategyUpdate): MessagingStrategyUpdate {
+    val new = rolloutStrategyUpdate.new
+    val old = rolloutStrategyUpdate.old
+    return MessagingStrategyUpdate()
+      .let { if (new != null) it.newStrategy(toMessagingRolloutStrategy(new)) else it }
+      .let { if (old != null) it.oldStrategy(toMessagingRolloutStrategy(old)) else it }
+      .updateType(StrategyUpdateType.fromValue(rolloutStrategyUpdate.type.uppercase())!!)
+  }
+
+  private fun toMessagingRolloutStrategy(rolloutStrategy: RolloutStrategy): MessagingRolloutStrategy {
+    val attributes = rolloutStrategy.attributes
+    return MessagingRolloutStrategy()
+      .id(rolloutStrategy.id!!)
+      .name(rolloutStrategy.name)
+      .percentage(rolloutStrategy.percentage)
+      .percentageAttributes(rolloutStrategy.percentageAttributes)
+      .value(rolloutStrategy.value)
+      .attributes(attributes?.map { rsa: RolloutStrategyAttribute -> toRolloutStrategyAttribute(rsa) } ?: ArrayList())
+  }
+
+  private fun toRolloutStrategyAttribute(rolloutStrategyAttribute: RolloutStrategyAttribute): MessagingRolloutStrategyAttribute {
+    return MessagingRolloutStrategyAttribute()
+      .conditional(rolloutStrategyAttribute.conditional)
+      .values(rolloutStrategyAttribute.values)
+      .fieldName(rolloutStrategyAttribute.fieldName)
+      .type(rolloutStrategyAttribute.type)
   }
 }
