--- backend/mr-messaging-processor/pom.xml
@@ -1,90 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>io.featurehub.messaging</groupId>
-  <artifactId>mr-messaging-processor</artifactId>
-  <version>1.1-SNAPSHOT</version>
-  <name>mr-messaging-processor</name>
-
-  <description>
-    This artifact simply deals with extracting the feature messaging update cloud events and dropping them into the cloud event
-    registry. It can be included into any end application that needs to listen to the stream
-  </description>
-
-  <url>https://featurehub.io</url>
-  <developers>
-    <developer>
-      <email>irina@featurehub.io</email>
-      <id>isouthwell</id>
-      <name>Irina Southwell</name>
-      <organization>Anyways Labs Ltd</organization>
-    </developer>
-
-    <developer>
-      <email>richard@featurehub.io</email>
-      <id>rvowles</id>
-      <name>Richard Vowles</name>
-      <organization>Anyways Labs Ltd</organization>
-    </developer>
-  </developers>
-
-  <licenses>
-    <license>
-      <name>Apache 2 with Commons Clause</name>
-      <url>https://github.com/featurehub-io/featurehub/blob/master/LICENSE.txt</url>
-    </license>
-  </licenses>
-
-  <dependencies>
-    <dependency>
-      <groupId>io.featurehub.composites</groupId>
-      <artifactId>composite-app</artifactId>
-      <version>1.1-SNAPSHOT</version>
-    </dependency>
-
-    <dependency>
-      <groupId>io.featurehub.api</groupId>
-      <artifactId>messaging-api</artifactId>
-      <version>1.1-SNAPSHOT</version>
-    </dependency>
-
-    <dependency>
-      <groupId>io.featurehub.messaging</groupId>
-      <artifactId>messaging-channel</artifactId>
-      <version>1.1-SNAPSHOT</version>
-    </dependency>
-
-    <dependency>
-      <groupId>io.featurehub.mr</groupId>
-      <artifactId>mr-db-services</artifactId>
-      <version>1.1-SNAPSHOT</version>
-    </dependency>
-
-    <dependency>
-      <groupId>io.featurehub.composites</groupId>
-      <artifactId>composite-test</artifactId>
-      <version>[1.1, 2)</version>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>io.repaint.maven</groupId>
-        <artifactId>tiles-maven-plugin</artifactId>
-        <version>2.32</version>
-        <extensions>true</extensions>
-        <configuration>
-          <filtering>false</filtering>
-          <tiles>
-            <tile>io.featurehub.tiles:tile-java:[1.1,2)</tile>
-          </tiles>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-
-
-</project>
